% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mobility.R
\name{assign_mobility}
\alias{assign_mobility}
\title{assign_mobility}
\usage{
assign_mobility(.sch_exp, .n_sch, .wt_vec, .wt_nonmob = FALSE, ...)
}
\arguments{
\item{.sch_exp}{A matrix or dataframe. The school-level information created
by the \code{\link{gen_u_mmrem}} function and expanded by the
\code{\link{expand_sch_info}} function.}

\item{.n_sch}{Numeric scalar. Gives the total number of schools in the
dataset. The variance-covariance matrix for predictor z will have dimensions
\code{.n_sch x .n_sch}.}

\item{.wt_vec}{A numeric vector with length equal to the maximum number
of schools attended by students in the data (in this simulation, the
maximum number is 2). The values in \code{.wt_vec} are used to weight
the effects of different schools attended on students. For this study,
all mobile students must have the same weights. If different weighting
patterns are desired, the code will need to be updated. Defaults to
\code{.wt_vec = rep(0.5, 2)}.}

\item{.wt_nonmob}{Logical. Indicates whether non-mobile students should
receive the same weights as mobile students. Technically, it shouldn't
matter if non-mobile students are given the same weights because their
first and second schools are the same, so all weighting schemes should
be equivalent, but it may matter for passing data to MLwiN for estimation.
See \code{\link[R2MLwiN]{runMLwiN}} for more information.}

\item{...}{Other parameters passed to \code{\link{simulate_mobility}}.}
}
\value{
This function returns a dataframe with a number of columns equal to
\code{NCOL(.sch_exp) * 2 + 3}. The columns from \code{.sch_exp} are repeated
for the second school, and the mobility assignments from
\code{\link{simulate_mobility}} along with the first and second school
weights are appended as the final 3 columns.
}
\description{
This function uses mobility assignments simulated by
\code{\link{simulate_mobility}} to identify a student's second school
information. If the student is not mobile, their second school will be
equal to their first school. If they are mobile, their second school ID will
be determined by their first school ID plus their mobility indicator (i.e.,
representing a change to an adjacent, or nearly adjacent, school). For
students attending the first school (i.e., ID = 1) or the last school
(i.e., ID = total number of schools), their second school, if they are
mobile, will be schools above (for ID = 1) or schools below (for ID = max
school ID).
}
\details{
Note: this function was designed for a simulation in which the maximum
number of schools attended by each student is equal to 2. The code would
need to be updated if it is desired that students could attend more than
2 schools.
}
\examples{
\dontrun{

library(magrittr)

sch_inf <-
  gen_u_mmrem(
    .n_sch = 50,
    clust_cov = c(0.8, 0.2, 0.1)
  ) \%>\%
  expand_sch(., .n_sch = 50, n_stu = 50) \%>\%
  assign_mobility(., .n_sch = 50, .wt_vec = c(0.5, 0.5))

}
}
