% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mm_list.R
\name{build_mm_list}
\alias{build_mm_list}
\title{build_mm_list}
\usage{
build_mm_list(
  .mm_ids = paste0("ids_", seq_len(.n_sch)),
  .mm_wts = paste0("wts_", seq_len(.n_sch))
)
}
\arguments{
\item{.mm_ids}{A character vector. The names of the multiple membership
unique ID variables.}

\item{.mm_wts}{A character vector. The names of the multiple membership
weight variables.}
}
\value{
This function returns a list to be passed to \code{run_mlwin}.
}
\description{
build_mm_list
}
\examples{
\dontrun{

}
}
