% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_icc.R
\name{calc_exp_icc}
\alias{calc_exp_icc}
\title{calc_exp_icc}
\usage{
calc_exp_icc(
  .pct_mobile = 0.2,
  .clust_cov = c(0.8, 0.1),
  .wt_vec = c(0.5, 0.5),
  .u_resid_var = 0.2,
  .gamma_z = 1,
  .gamma_x = c(10, 1.5),
  .var_x = 4,
  .var_r = 2
)
}
\arguments{
\item{.pct_mobile}{A numeric scalar between 0 and 1. Gives the proportion of
mobile students to be saved in the "center" option of
\code{.mob_option}.}

\item{.clust_cov}{Numeric vector. The first element of the vector gives the
variance of all schools' predictors, z. If present, the second element gives
the covariance of z between schools k and k + 1. The values given in
\code{.clust_cov} apply to all schools (that is, similar to a Toeplitz
pattern). Any off-diagonal values (i.e., covariances) not specified will
default to 0. The main diagonal (i.e., predictor variance) defaults to 0.8.}

\item{.gamma_x}{Numeric vector with length p (where p is the number of model
coefficients, including the intercept).}
}
\value{
A numeric scalar.
}
\description{
calc_exp_icc
}
\examples{
\dontrun{

}
}
