% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_l2var.R
\name{calc_exp_l2var}
\alias{calc_exp_l2var}
\title{calc_exp_l2var}
\usage{
calc_exp_l2var(
  .pct_mobile = 0,
  .clust_cov = c(0.8, 1),
  .wt_vec = c(0.5, 0.5),
  .u_resid_var = 0.2,
  .gamma_z = 1
)
}
\arguments{
\item{.pct_mobile}{A numeric scalar between 0 and 1. Gives the proportion of
mobile students to be saved in the "center" option of
\code{.mob_option}.}

\item{.clust_cov}{Numeric vector. The first element of the vector gives the
variance of all schools' predictors, z. If present, the second element gives
the covariance of z between schools k and k + 1. The values given in
\code{.clust_cov} apply to all schools (that is, similar to a Toeplitz
pattern). Any off-diagonal values (i.e., covariances) not specified will
default to 0. The main diagonal (i.e., predictor variance) defaults to 0.8.}
}
\value{
A numeric scalar.
}
\description{
calc_exp_l2var
}
\examples{
\dontrun{

}
}
