% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_multicol.R
\name{collapse_vals}
\alias{collapse_vals}
\title{collapse_vals}
\usage{
collapse_vals(..., .exclude_val = 0)
}
\arguments{
\item{...}{Values passed to the internal aggregator functions.}
}
\description{
Helper function for pivoting wide & then summarizing/aggregating in the
\code{\link{pivot_wider_multicol}} steps. \code{\link{collapse_vals}}
aggregates using \code{\link[base]{unique}}, then removing unwanted values.
If the exclusion procedure results in no values being retained, the original
values are passed to \code{\link{combine_vals}}.
}
\seealso{
\code{\link{combine_vals}}
}
