% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_multicol.R
\name{combine_vals}
\alias{combine_vals}
\title{combine_vals}
\usage{
combine_vals(...)
}
\arguments{
\item{...}{Values passed to the internal aggregator functions.}
}
\description{
Helper function for pivoting wide & then summarizing/aggregating in the
\code{\link{pivot_wider_multicol}} steps. \code{\link{combine_vals}}
aggregates using \code{\link[base]{sum}} with the key alteration that
\code{na.rm = TRUE}.
}
\seealso{
\code{\link{collapse_vals}}
}
