% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_sch.R
\name{expand_sch}
\alias{expand_sch}
\title{expand_sch}
\usage{
expand_sch(.sch_dat, .n_stu, .n_sch)
}
\arguments{
\item{.sch_dat}{A matrix or dataframe. The school-level information created
by the \code{\link{gen_u_mmrem}} function.}

\item{.n_stu}{A numeric scalar. The number of students attending each
school. Note: this is not the total number of students in the dataset,
merely the number of students per school.}

\item{.n_sch}{Numeric scalar. Gives the total number of schools in the
dataset. The variance-covariance matrix for predictor z will have dimensions
\code{.n_sch x .n_sch}.}
}
\value{
This function returns a dataframe with a number of rows equal to
\code{.n_stu * .n_sch}. The columns of the input data (\code{.sch_dat}) are
retained in the output.
}
\description{
This function expands school-level information generated by
\code{\link{gen_u_mmrem}}, repeating each school's info a number of times
equal to the number of students attending each school. That is, if there
are \code{.n_stu} per school and \code{.n_sch}, the final output should have
\code{.n_stu * .n_sch} rows.
}
\examples{
\dontrun{

# generate data using the gen_u_mmrem function
dat <- gen_u_mmrem(.n_sch = 10)

# expand the data to repeat school information for each student
# attending a given school
expand_sch(dat, .n_stu = 10, .n_sch = 10)

}
}
