% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_weights.R
\name{expand_weights}
\alias{expand_weights}
\title{expand_weights}
\usage{
expand_weights(
  .dat,
  .id_cols,
  .cols,
  .wider_names,
  .wider_values,
  .wider_prefix = "wts_",
  .tag_name = "time",
  ...
)
}
\arguments{
\item{.dat}{A dataframe containing the data to be pivoted.}

\item{.id_cols}{A string or \code{\link[tidyselect]{starts_with}} selection
helper that identifies row ID.}

\item{.cols}{Either a string vector listing the names of the columns to be
pivoted or a \code{\link[tidyselect]{starts_with}} selection helper.}

\item{.wider_names}{A string. Identifies the "tagless" variable that will
be used to construct names for the new expanded weight variables. For
example, if the original, wide-format variable is "sch_id_1", the "tagless"
variable will be "sch_id". Ultimately, this "tagless" name will be
determined by the regex given by \code{.capture_groups} passed to
\code{\link{pivot_longer_multicol}}.}

\item{.wider_values}{A string. Identifies the "tagless" variable that will
be used to add values to the new expanded weight variables. For
example, if the original, wide-format variable is "sch_wt_1", the "tagless"
variable will be "sch_wt". Ultimately, this "tagless" name will be
determined by the regex given by \code{.capture_groups} passed to
\code{\link{pivot_longer_multicol}}.}

\item{.tag_name}{String. Indicates the variable name that should be given to
the new variable containing the variable tags. Because this function was
created to manage the longitudinal school mobility process, the argument
defaults to "time".}

\item{...}{Other parameters passed to \code{\link{pivot_longer_multicol}}.}
}
\value{
This function outputs a dataframe with the same number of rows as
the original input, but with only the new columns created by expanding
the weight variables plus the ID columns given in \code{.id_cols}.
The new dataframe should have a number of columns
equal to the unique values of \code{.wider_names}.
}
\description{
This function uses a combination of \code{\link{pivot_longer_multicol}},
\code{\link[tidyr]{pivot_wider}}, and \code{\link[dplyr]{group_by}} with
\code{\link[dplyr]{summarise}} to expand the weights for each school
a student attended to a single column per school ID.
}
\examples{
\dontrun{

# construct fake data
temp_dat <-
  tibble::tibble(
    x = rnorm(100),
    y = rnorm(100),
    z_1 = rep(c(1, 0), 50),
    z_2 = rep(c(5, 6, 7, 8), 25),
    g_1 = rnorm(100),
    g_2 = rnorm(100)
  )

temp_dat \%>\%
  expand_weights(
    .dat = .,
    .id_cols = "x",
    .cols = tidyr::matches("_"),
    .wider_names = "z",
    .wider_values = "g",
    .wider_prefix = "new_"
  )

}
}
