% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sumstat.R
\name{extract_sumstat}
\alias{extract_sumstat}
\title{extract_sumstat}
\usage{
extract_sumstat(.mod)
}
\arguments{
\item{.mod}{A model object produced by \code{\link[R2MLwiN]{runMLwiN}}.}
}
\value{
This function returns a single-column dataframe with model fit
estimates (DIC).
}
\description{
extract_sumstat
}
\examples{
\dontrun{


}
}
