% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_off_diag.R
\name{is_off_diag}
\alias{is_off_diag}
\title{is_off_diag}
\usage{
is_off_diag(.dat, .off_diag = 1, sub_diag = TRUE)
}
\arguments{
\item{.dat}{A matrix or dataframe.}

\item{.off_diag}{A numeric scalar. \code{NROW(.dat) - .off_diag} must
be greater than or equal to 1. If a negative value is entered, the absolute
value will be taken. If \code{.off_diag = 0}, the primary matrix diagonal is
identified. Defaults to 1.}

\item{sub_diag}{Logical. Indicates whether the sub-diagonal (lower
off-diagonal) or the super-diagonal (upper off-diagonal) should be
identified. If \code{sub_diag = TRUE}, the sub-diagonal will be identified,
if \code{sub_diag = FALSE}, the super-diagonal will be returned instead.
Defaults to \code{TRUE}.}
}
\value{
This function returns a logical matrix with the same dimensions as
\code{.dat}. All cells but those lying on the desired off-diagonal will take
the value of \code{FALSE}, while those on the off-diagonal will return
a value of \code{TRUE}.
}
\description{
This function tests an input matrix and returns a logical matrix of the
same dimensions, with \code{TRUE} indicating that a cell is in the desired
off-diagonal. By default, the function returns the sub-diagonal (i.e.,
the lower off-diagonal).
}
\examples{
\dontrun{

nr <- 4
nc <- 6

mat <- matrix(seq(4*6), nrow = nr, ncol = nc)
df <- as.data.frame(mat)

# works for matrices
is_off_diag(mat, .off_diag = 2)
is_off_diag(mat, .off_diag = 2, sub_diag = F)

# and dataframes
is_off_diag(df, .off_diag = 2)
is_off_diag(df, .off_diag = 2, sub_diag = F)

# check for equality
all(is_off_diag(mat, 2) == is_off_diag(df, 2))
all(is_off_diag(mat, 2, sub_diag = F) == is_off_diag(df, 2, sub_diag = F))

}
}
