% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_conditions_funs.R
\name{list_conditions}
\alias{list_conditions}
\title{list_conditions}
\usage{
list_conditions()
}
\value{
A named list of all combinations of arguments given in the parent
function. The list has length equal to the number of arguments passed to
the function and each argument has length equal to the number of
combinations.
}
\description{
This function belongs in the interior of
\code{\link{list_conditions_datagen}} and
\code{\link{list_conditions_estimate}}. It takes the arguments passed to
its parent function and returns a list of all combinations of those
arguments.
}
\details{
The argument combinations are created using \code{\link[purrr]{cross}} to
create a list where each element is a combination (list length is equal
to the number of combinations), and then the list is transformed using
\code{\link[purrr]{transpose}} to flip it "inside out" such that the
result has length equal to the number of arguments passed to the parent
function, and each argument has length equal to the number of combinations.
}
