% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mlwin.R
\name{run_mlwin}
\alias{run_mlwin}
\title{run_mlwin}
\usage{
run_mlwin(
  .dat,
  .model_formula = as.formula("y ~ 1 + (1 | sch_id_1) + (1 | stu_id)"),
  .outcome_dist = "Normal",
  .mcmc_burn = 100,
  .mcmc_iter = 1000,
  .mcmc_nchains = 1,
  .mm_list = NULL
)
}
\arguments{
\item{.dat}{A dataframe produced by \code{\link{generate_data}}.}

\item{.model_formula}{A formula object.}

\item{.outcome_dist}{A string. Defaults to "Normal".}

\item{.mcmc_burn}{Numeric. Defaults to 100.}

\item{.mcmc_iter}{Numeric. Defaults to 1000.}

\item{.mcmc_nchains}{Numeric. Defaults to 1.}

\item{.mm_list}{A list. Generated by \code{\link{build_mm_list}}}
}
\value{
Returns a model object produced by \code{\link[R2MLwiN]{runMLwiN}}.
}
\description{
run_mlwin
}
\examples{
\dontrun{

}
}
