% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{run_sim}
\usage{
run_sim(
  .sim_fun = generate_data,
  .sim_args = list_conditions_datagen(),
  .plan_future = plan_future(),
  .seed_start = TRUE,
  ...
)
}
\arguments{
\item{.sim_fun}{A function name. The function to be iterated over in the
simulations.}

\item{.sim_args}{A named list of simulation conditions. Each element must
match a named argument in \code{.sim_fun}, and each element must have
length equal to the number of combinations of conditions. Alternately,
pass a call to \code{\link{list_conditions_datagen}} or
\code{\link{list_conditions_estimate}}.}

\item{.plan_future}{A call to \code{\link{plan_future}}.}

\item{.seed_start}{A logical or numeric. Passed to the \code{seed}
argument of \code{\link[furrr]{furrr_options}}. If \code{TRUE}, \code{furrr}
will use L'Ecuyer-CMRG RNG streams for parallel-safe random number generation
that will produce identical results regardless of the parallel strategy
set by \code{.plan_future}. If \code{.seed_start} is a numeric scalar,
all runs of the \code{run_sim} function with identical \code{.sim_fun} and
\code{.sim_args} will produce identical results.}

\item{...}{Parameters passed to \code{sim_fun}.}
}
\value{
A list of objects created by applying \code{.sim_fun} to each
combination of conditions given in \code{.sim_args}.
}
\description{
run_sim
}
\examples{
\dontrun{

.n_reps <- 1:2

purrr::map(.n_reps, ~run_sim(
  .sim_args = list_conditions_datagen(
    .n_sch = 5,
    .n_stu = 5,
    .clust_cov = .8,
    .pct_mobile = .5,
    .gamma_x = list(c(0, 0))
   )
 )) \%>\%
 rlang::set_names(., nm = paste0("replication_", .n_reps))

}
}
