% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_datagen_funs.R
\name{run_sim_datagen}
\alias{run_sim_datagen}
\title{run_sim_datagen}
\usage{
run_sim_datagen(
  .n_reps = 1,
  .seed_start = 33,
  .n_cores = future::availableCores() - 1,
  .plan = future::sequential,
  .n_sch = 50,
  .n_stu = 50,
  .u_resid_var = 0.2,
  .clust_cov = list(c(0.8, 0), c(0.8, 0.25), c(0.8, 0.4)),
  .wt_vec = list(c(0.5, 0.5)),
  .pct_mobile = list(0, 0.25, 0.5),
  .mean_x = 5,
  .var_x = 4,
  .mean_r = 0,
  .var_r = 2,
  .gamma_z = 1,
  .gamma_x = list(c(10, sqrt(17)/2), c(10, sqrt(11/3)/2), c(10, 1/(2 * sqrt(3))))
)
}
\arguments{
\item{.n_reps}{Numeric. Either the total number of replications to run or
a vector indicating the replications to run. Defaults to 1.}

\item{.seed_start}{Numeric scalar. The starting seed value. Each function
in this script that uses random number generation will take this starting
seed and add some noise to it. Defaults to NULL.}

\item{.n_cores}{Numeric scalar or call to
\code{\link[future]{availableCores}}.}

\item{.plan}{A call to \code{\link[future]{multisession}} or
\code{\link[future]{sequential}}l}

\item{.n_sch}{Numeric scalar. Gives the total number of schools in the
dataset. The variance-covariance matrix for predictor z will have dimensions
\code{.n_sch x .n_sch}.}

\item{.n_stu}{A numeric scalar. The number of students attending each
school. Note: this is not the total number of students in the dataset,
merely the number of students per school.}

\item{.clust_cov}{Numeric vector. The first element of the vector gives the
variance of all schools' predictors, z. If present, the second element gives
the covariance of z between schools k and k + 1. The values given in
\code{.clust_cov} apply to all schools (that is, similar to a Toeplitz
pattern). Any off-diagonal values (i.e., covariances) not specified will
default to 0. The main diagonal (i.e., predictor variance) defaults to 0.8.}

\item{.pct_mobile}{A numeric scalar between 0 and 1. Gives the proportion of
mobile students to be saved in the "center" option of
\code{.mob_option}.}

\item{.gamma_x}{Numeric vector with length p (where p is the number of model
coefficients, including the intercept).}
}
\value{
This function returns a list of length equal to \code{.n_reps}. The
second level of the list has length equal to the product of the number of
levels of each argument.
}
\description{
This function is a wrapper around \code{\link{generate_data}}. Each of
the arguments (except \code{.n_reps}) take the values for the conditions.
Fixed conditions with length 1 can be input as scalar values. Variable
conditions and fixed conditions with length > 1 must be entered as lists
with each list element representing the values of the argument for a single
condition. All arguments must have length 1 or the length of the input with
the max value. For simulations that have different numbers of levels
within factors, use \code{\link[base]{rep}} to repeat the values to the
appropriate size. For example: \code{.pct_mobile = rep(list(0, .25, .5), 2)}.
}
\examples{
\dontrun{

run_sim_datagen(
  .n_reps = 1,
  .n_sch = 50,
  .n_stu = 50,
  .u_resid_var = 0.2,
  .clust_cov = rep(list(
    c(0.8, 0),
    c(0.8, 0.25)
  ), 3),
  .wt_vec = list(c(0.5, 0.5)),
  .pct_mobile = rep(list(0, 0.25, 0.5), 2),
  .mean_x = 5,
  .var_x = 4,
  .mean_r = 0,
  .var_r = 2,
  .gamma_z = 1,
  .gamma_x = rep(list(
    c(10, sqrt(17)/2),
    c(10, sqrt(11/3)/2),
    c(10, 1/(2*sqrt(3)))
  ), 2)
)


}
}
