% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_datagen_funs.R
\name{run_sim_datagen_inner}
\alias{run_sim_datagen_inner}
\title{run_sim_datagen_inner}
\usage{
run_sim_datagen_inner(
  .n_sch,
  .n_stu,
  .u_resid_var,
  .clust_cov,
  .wt_vec,
  .pct_mobile,
  .mean_x,
  .var_x,
  .mean_r,
  .var_r,
  .gamma_z,
  .gamma_x,
  .cond_index,
  .rep_index,
  ...
)
}
\arguments{
\item{.n_sch}{Numeric scalar. Gives the total number of schools in the
dataset. The variance-covariance matrix for predictor z will have dimensions
\code{.n_sch x .n_sch}.}

\item{.n_stu}{A numeric scalar. The number of students attending each
school. Note: this is not the total number of students in the dataset,
merely the number of students per school.}

\item{.clust_cov}{Numeric vector. The first element of the vector gives the
variance of all schools' predictors, z. If present, the second element gives
the covariance of z between schools k and k + 1. The values given in
\code{.clust_cov} apply to all schools (that is, similar to a Toeplitz
pattern). Any off-diagonal values (i.e., covariances) not specified will
default to 0. The main diagonal (i.e., predictor variance) defaults to 0.8.}

\item{.pct_mobile}{A numeric scalar between 0 and 1. Gives the proportion of
mobile students to be saved in the "center" option of
\code{.mob_option}.}

\item{.gamma_x}{Numeric vector with length p (where p is the number of model
coefficients, including the intercept).}

\item{.rep_index}{Numeric. A unique number identifying the current
data gen replication number. Created from within
\code{\link{run_sim_datagen}}.}

\item{...}{Other values from the arguments to \code{\link[purrr]{map}} that
are passed to the inner loop of \code{\link{run_sim_datagen}} but not
used in the inner loop function.}
}
\value{
This function returns a list of the current condition specifications
and the data generated from it.
}
\description{
This is the innermost loop called from \code{\link{run_sim_datagen}} that
loops over the data generation conditions.
}
