% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_datagen_funs.R
\name{run_sim_datagen_outer}
\alias{run_sim_datagen_outer}
\title{run_sim_datagen_outer}
\usage{
run_sim_datagen_outer(
  .rep_index,
  .sim_args,
  .n_reps,
  .n_cond,
  .prog_reps,
  .start_time,
  ...
)
}
\arguments{
\item{.rep_index}{Numeric. A unique number identifying the current
data gen replication number. Created from within
\code{\link{run_sim_datagen}}.}

\item{.sim_args}{A named list of condition values, created in
\code{\link{run_sim_datagen}}.}

\item{.n_reps}{Numeric scalar. The total number of replications to run.
Defaults to 1.}

\item{.n_cond}{Numeric. The total number of unique conditions for data
generation. Created in \code{\link{run_sim_datagen}}.}

\item{.prog_reps}{An object of class "progress_bar". Created by the
\code{\link[progress]{progress_bar}} function in
\code{\link{run_sim_datagen}}.}

\item{.start_time}{Numeric. The elapsed time since the simulation function
started (given by \code{\link[tictoc]{tic}}).}

\item{...}{Other values from the arguments to \code{\link[purrr]{map}} that
are passed to the outer loop of \code{\link{run_sim_datagen}} but not
used in the outer loop function.}
}
\value{
This function returns the data generated over all conditions and
replications as well as the meta data (e.g., condition information and
replication numbers).
}
\description{
This is the outermost loop called from \code{\link{run_sim_datagen}} that
loops over the simulation condition replications.
}
