% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mobility_TEMP.R
\name{simulate_mobility_TEST}
\alias{simulate_mobility_TEST}
\title{simulate_mobility_TEST}
\usage{
simulate_mobility_TEST(.n_sch, .n_stu)
}
\arguments{
\item{.n_sch}{}
}
\value{
Matrix.
}
\description{
Testing out a new way to simulate mobility.
}
\examples{
\dontrun{



}
}
