% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_formula.R
\name{write_formula}
\alias{write_formula}
\title{write_formula}
\usage{
write_formula(
  response_var = "y",
  predictor_var = NULL,
  level1_id = "stu_id",
  level2_id1 = "sch_id_1",
  level3_id1 = NULL,
  random_l1_var = NULL,
  random_l2_var = NULL,
  random_l3_var = NULL,
  intercept = "1"
)
}
\arguments{
\item{response_var}{String. Gives the name of the response variable. Note:
if the response variable is not normally distributed, the link must be
included (see \code{\link[R2MLwiN]{runMLwiN}}) for details. Also note: for
non-normal links, a constant variable (given as "cons" or "denom" in the
\code{R2MLwiN}) documentation must be added to the data and included in the
link. For example: \code{"logit(outcome, cons)"}.}

\item{predictor_var}{A vector of strings. Gives the names of the predictor
variables (fixed effects). For an unconditional model, set
\code{predictor_var = NULL}.}

\item{level1_id}{A string. Gives the name of the level-1 id variable.}

\item{level2_id1}{A string. Gives the name of the first level-2 id variable.}

\item{level3_id1}{A string. Gives the name of the first level-3 id variable.}

\item{random_l1_var}{A vector of strings. Gives the names of the variables
to be used to explain complex level-1 variation.}

\item{random_l2_var}{A vector of strings. Gives the names of the variables
to be used to explain complex level-2 variation.}

\item{random_l3_var}{A vector of strings. Gives the names of the variables
to be used to explain complex level-3 variation.}

\item{intercept}{String. Indicates the value of the intercept. The intercept
defaults to 1. As of 2-10-21, \code{\link[R2MLwiN]{runMLwiN}} fails when
a non-1 intercept is given. For mixed distribution or multi-category
outcomes, a constant intercept can be specified using brackets:
\code{"1[1]"}.}
}
\value{
A formula object.
}
\description{
write_formula
}
\examples{
\dontrun{

write_formula("y", c("x_predictor", "z_composite"))

}
}
