% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mm_list.R
\name{build_mm_list}
\alias{build_mm_list}
\title{build_mm_list}
\usage{
build_mm_list(.n_clusters, .mm_id_nms = "ids", .mm_wt_nms = "wts")
}
\arguments{
\item{.n_clusters}{Numeric scalar. The number of clusters in the data.}

\item{.mm_id_nms}{A string. The prefix name of the multiple membership
unique ID variables.}

\item{.mm_wt_nms}{A string. The prefix name of the multiple membership
weight variables.}
}
\value{
This function returns a list to be passed to \code{run_mlwin}.
}
\description{
build_mm_list
}
\examples{
\dontrun{

}
}
