% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mm_list.R
\name{build_mm_list}
\alias{build_mm_list}
\title{build_mm_list}
\usage{
build_mm_list(.n_clust, .mm_id_nms = "sch_id_", .mm_wt_nms = "sch_wt_")
}
\arguments{
\item{.n_clust}{Numeric scalar. The number of clusters.
Depending on whether the "compact" (default) or "wide" format was used in
to \code{\link{generate_data}}, this argument either refers to the total
number of clusters in the dataset ("wide"), or the maximum number of clusters
an individual can be a member of ("compact"). For example, if students
attended a maximum of 2 schools, but there were 50 schools in the dataset,
for the "wide" option, set \code{.n_clust = 50}; for the "compact"
option, set \code{.n_clust = 2}.}

\item{.mm_id_nms}{A string. The prefix name of the multiple membership
unique ID variables.}

\item{.mm_wt_nms}{A string. The prefix name of the multiple membership
weight variables.}
}
\value{
This function returns a list to be passed to \code{run_mlwin}.
}
\description{
build_mm_list
}
\examples{
\dontrun{

}
}
