% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_groups.R
\name{extract_groups}
\alias{extract_groups}
\title{extract_groups}
\usage{
extract_groups(.strings, .capture_groups = "(.+)_(.+)", .tag_group = 2)
}
\arguments{
\item{.strings}{A character vector. Gives the strings to be extracted by
\code{.capture_groups}.}
}
\value{
This function returns a character vector that is the result of
matching the input \code{.strings} against the regex \code{.capture_groups},
selecting the desired capture group, \code{.tag_groups}, and returning
the unique values of the result.
}
\description{
This function takes a string input and applies the
\code{\link[stringr]{str_match}} function to extract the unique values
of the capture group requested by \code{.tag_group} based on a regex
given by \code{.capture_groups}.
}
\examples{
\dontrun{

char_vec <- c("extract", "a_b_c_d", "1_m.b-q_m")

# accept 1 or more matches in each capture group
extract_groups(char_vec, "(.+)_(.+)", 1)
extract_groups(char_vec, "(.+)_(.+)", 2)

# accept 0 or 1 matches in each capture group
extract_groups(char_vec, "(.?)a(.?)", 1)
extract_groups(char_vec, "(.?)a(.?)", 2)


}
}
