% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(
  .n_sch = 5,
  .n_stu = 5,
  .u_resid_var = 0.2,
  .clust_cov = c(0.8, 1),
  .wt_vec = c(0.5, 0.5),
  .pct_mobile = 0,
  .mean_x = 5,
  .var_x = 4,
  .mean_r = 0,
  .var_r = 2,
  .gamma_z = 0,
  .gamma_x = c(10, sqrt(17)/2),
  .seed_start = NULL,
  ...
)
}
\arguments{
\item{.n_sch}{Numeric scalar. Gives the total number of schools in the
dataset. The variance-covariance matrix for predictor z will have dimensions
\code{.n_sch x .n_sch}.}

\item{.n_stu}{A numeric scalar. The number of students attending each
school. Note: this is not the total number of students in the dataset,
merely the number of students per school.}

\item{.clust_cov}{Numeric vector. The first element of the vector gives the
variance of all schools' predictors, z. If present, the second element gives
the covariance of z between schools k and k + 1. The values given in
\code{.clust_cov} apply to all schools (that is, similar to a Toeplitz
pattern). Any off-diagonal values (i.e., covariances) not specified will
default to 0. The main diagonal (i.e., predictor variance) defaults to 0.8.}

\item{.pct_mobile}{A numeric scalar between 0 and 1. Gives the proportion of
mobile students to be saved in the "center" option of
\code{.mob_option}.}

\item{.gamma_x}{Numeric vector with length p (where p is the number of model
coefficients, including the intercept).}

\item{.seed_start}{Numeric scalar. The starting seed value. Each function
in this script that uses random number generation will take this starting
seed and add some noise to it. Defaults to NULL.}

\item{...}{Other parameters passed to \code{\link{assign_mobility}}.}
}
\value{
This function returns data generated under the correlated cluster
multiple membership model, with correlated clusters defined by the
correlation between the school-level predictor z. The result is a tibble
with number of rows equal to \code{.n_sch * .n_stu} and a number of columns
equal to \code{18 + .n_sch}.
}
\description{
This function
}
\examples{
\dontrun{

# with the following defaults, we adjust x1beta to get the
# desired icc:

## icc = 0.05
# x1beta <- sqrt(17)/2

## icc = 0.15
# x1beta <- sqrt(11/3)/2

## icc = 0.30
x1beta <- 1/(2*sqrt(3))

generate_data(
  .n_sch = 5,
  .n_stu = 5,
  .u_resid_var = 0.2,
  .clust_cov = c(.8, 1),
  .wt_vec = c(0.5, 0.5),
  .pct_mobile = 0,
  .mean_x = 5,
  .var_x = 4,
  .mean_r = 0,
  .var_r = 2,
  .gamma_z = 0,
  .gamma_x = c(10, x1beta),
  .seed_start = 1
)

}
}
